+++ anchor/programs/depo/src/constants.rs
@@ -1,0 +1,1 @@
+pub const SOL_TO_LAMPORTS: u64 = 1_000_000_000;
