--- anchor/programs/depo/src/lib.rs
@@ -1,12 +1,0 @@
-#![allow(clippy::result_large_err)]
-
-use anchor_lang::prelude::*;
-
-declare_id!("coUnmi3oBUtwtd9fjeAvSsJssXh5A5xyPbhpewyzRVF");
-
-#[program]
-pub mod depo {
-
-  // TODO
-
-}
