+++ anchor/programs/depo/src/state/modules/mod.rs
@@ -1,0 +1,15 @@
+pub mod module_type;
+pub mod expiry_fallback;
+pub mod minimum_amount;
+pub mod multisig_approval;
+pub mod single_approval;
+pub mod target_amount;
+pub mod timelock;
+
+pub use module_type::*;
+pub use expiry_fallback::*;
+pub use minimum_amount::*;
+pub use multisig_approval::*;
+pub use single_approval::*;
+pub use target_amount::*;
+pub use timelock::*;
