+++ anchor/programs/depo/src/state/modules/module_type.rs
@@ -1,0 +1,11 @@
+use anchor_lang::prelude::*;
+
+#[derive(InitSpace, AnchorSerialize, AnchorDeserialize, Clone)]
+pub enum ModuleType {
+    TimeLock,
+    ExpiryFallback,
+    SingleApproval,
+    MultisigApproval,
+    TargetAmount,
+    MinimumAmount
+}
