+++ anchor/programs/depo/src/state/recipient.rs
@@ -1,0 +1,10 @@
+use anchor_lang::prelude::*;
+
+#[account]
+#[derive(InitSpace)]
+pub struct Recipient {
+  pub escrow: Pubkey,
+  pub wallet: Pubkey,
+  pub amount: u64,
+  pub has_withdrawn: bool
+}
