--- anchor/programs/deposit/src/accounts/create_escrow.rs
@@ -1,1 +1,0 @@
-use anchor_lang::prelude::*;
