--- anchor/programs/deposit/src/accounts/mod.rs
@@ -1,1 +1,0 @@
-pub mod create_escrow;
