--- anchor/programs/deposit/src/constants.rs
@@ -1,1 +1,0 @@
-pub const SOL_TO_LAMPORTS: u64 = 1_000_000_000;
