--- anchor/programs/deposit/src/instructions/mod.rs
@@ -1,1 +1,0 @@
-pub mod create_escrow;
