--- anchor/programs/deposit/src/state/mod.rs
@@ -1,9 +1,0 @@
-pub mod escrow;
-pub mod depositor;
-pub mod recipient;
-pub mod modules;
-
-pub use escrow::*;
-pub use depositor::*;
-pub use recipient::*;
-pub use modules::*;
