--- anchor/programs/deposit/src/state/modules/expiry_fallback.rs
@@ -1,13 +1,0 @@
-use anchor_lang::prelude::*;
-
-#[account]
-#[derive(InitSpace)]
-pub struct ExpiryFallback {
-  pub refund_after: u64
-}
-
-impl ExpiryFallback {
-  pub fn is_satisfied() {
-    // TODO: Implement the logic to check if satisfied
-  }
-}
