--- anchor/programs/deposit/src/state/modules/minimum_amount.rs
@@ -1,13 +1,0 @@
-use anchor_lang::prelude::*;
-
-#[account]
-#[derive(InitSpace)]
-pub struct MinimumAmount {
-  pub min_amount: u64   
-}
-
-impl MinimumAmount {
-  pub fn is_satisfied() {
-    // TODO: Implement the logic to check if satisfied
-  }
-}
