--- anchor/programs/deposit/src/state/modules/module_type.rs
@@ -1,11 +1,0 @@
-use anchor_lang::prelude::*;
-
-#[derive(InitSpace, AnchorSerialize, AnchorDeserialize, Clone)]
-pub enum ModuleType {
-    TimeLock,
-    ExpiryFallback,
-    SingleApproval,
-    MultisigApproval,
-    TargetAmount,
-    MinimumAmount
-}
